adding Authentication middleware

+++ routes/api.php
@@ -4,6 +4,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\api\PostController;
 use App\Http\Controllers\api\UserController;
+use App\Models\User;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\ValidationException;
 
 
 /*
@@ -21,10 +24,26 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::get('/posts',[PostController::class, "index"]);
-Route::post('/posts',[PostController::class, "store"]);
-Route::get('/posts/{id}',[PostController::class, "show"]);
-Route::put('/posts/{id}',[PostController::class, "update"]);
-Route::delete('/posts/{id}',[PostController::class, "destroy"]);
-Route::get('/users',[UserController::class, "index"]);
+Route::get('/posts',[PostController::class, "index"])->middleware("auth:sanctum");
+Route::post('/posts',[PostController::class, "store"])->middleware("auth:sanctum");
+Route::get('/posts/{id}',[PostController::class, "show"])->middleware("auth:sanctum");
+Route::put('/posts/{id}',[PostController::class, "update"])->middleware("auth:sanctum");
+Route::delete('/posts/{id}',[PostController::class, "destroy"])->middleware("auth:sanctum");
+Route::get('/users',[UserController::class, "index"])->middleware("auth:sanctum");
 
+Route::post('/sanctum/token', function (Request $request) {
+    $request->validate([
+        'email' => 'required|email',
+        'password' => 'required',
+        'device_name' => 'required',
+    ]);
+
+    $user = User::where('email', $request->email)->first();
+    if (! $user || ! Hash::check($request->password, $user->password)) {
+        throw ValidationException::withMessages([
+            'email' => ['The provided credentials are incorrect.'],
+        ]);
+    }
+
+    return $user->createToken($request->device_name)->plainTextToken;
+});
